ComparingAges.py
So, copy this code, go to Visual studio or any editor you use and paste it there. Run it and input all the values it asks you to. I know it would have been easier if I used loops but I wanted to make the code so that the beginners reading it could understand. Try it, and read the output in the terminal and let me know.